Change wrapPassthrough function.
This is currently a header function which is a detail of HIDL.
All uses of it currently have the same static casts around it.
By changing the return type to sp<IType>, the purpose of the
function becomes more clear, and there is less of a chance
false assumptions will be used by the (internal HIDL) users
of this function.

Bug: 67104214
Test: hidl's run_all_device_tests.sh
Change-Id: I9315d1b845bafebb2b3abfbf77a6b8d0c71e5cc2