tryShortenProcessName: for all threads.
tryShortenProcessName previously only applied to the main
thread. This CL applies it to all of the threads that have
been started.

It takes all threads with comm which are the first part of
the hidl package name (so, if the package is android.hardware,
it will rename all threads which are named 'android.hardwar').
This, for instance, won't rename threadpool threads (which
are given a name like HwBinder:806_1 from hwbinder). Other
threads which are manually given a name are also not renamed.

Process which are already given another name (don't start
with the HIDL package) are not renamed.

Test: boot on walleye
Test: see renamed threads in output of 'ps -AT'
Fixes: 77654200

Change-Id: Ia9c1008a02de275e3527f4c9b66e943e62b8030f