Mark android.hidl.token@1.0-utils as double_loadable
android.hidl.token@1.0-utils is explicitly marked as double_loadable
since it is dependended by another double_loadable library libgui.

Note: even without this change, the library is already capable of being
double loaded due to the dependency graph around it. This change is to
make it explicit so that double loading of a library is carefully
tracked and signed-off by the owner of the lib.

Bug: 77155589
Test: m -j
Change-Id: Ide2a93c768d169493e60b6d138842cec62bc8e24