Checks for grave getOrCreateCachedBinder problems.
toBinder failing caused b/69180318, and logs when it
fails would have made fixing the bug way easier. I've
made them fatal here since any failure indicates an
inconsistent state in the loaded libraries.

For the first (failure if entry is missing from the map),
if you are creating an object and sending it over binder,
you should have loaded the HIDL library in your process.

For the second (failure if calling the map function
doesn't return), this should only occur when an alloc in
the function fails which can also not happen in Android.

Fixes: 69408124
Test: boot Pixel 2 and ensure that 'getOrCreateCachedBinder'
    doesn't exist in the logs as expected

Change-Id: I330f3f2dfd04651457115bcb445761ef168317c7