MQDescriptor: allow copies.
When scatter-gather is used in order to read an FMQ object from a
binder buffer, we are left with a 'const MQDescriptor<...>*' which
points into the binder buffer. When this descriptor is part of a struct
which is returned, and this struct is returned using a copy (not
scatter gather since the struct contains a non-fixed-width item like
an interface), it must be copied into the default constructed member of
that struct.

Change-Id: Iafb3ecafc2822e90662d49e1d3d53b12ffde650b
Fixes: 117239572
Test: hidl_test